+++ R/simulator.R
@@ -5,7 +5,7 @@ NULL
 
 #' simulator
 #' 
-#' @param Nind int. Number of individuals
+#' @param grid int.  Number of cell per side of the environmental matrix
 #' @param Ngen int. Number of generations
 #' @param muG double. Mean of genetic values
 #' @param sigmaG double. Variance of genetic values
@@ -25,22 +25,30 @@ NULL
 #' simulator()
 #' 
 simulator <- function(
-  Nind = 50,
+  grid = 20,
   Ngen = 50,
   muG = 0,
   sigmaG = 1,
   muE = 0,
   sigmaE = 1,
-  Elim = 10,
+  Elim = 5,
   seedlings = 4,
   dispersal = 1,
   viability_deterministic = TRUE
 ){
-  sim <- simulatorCpp(Nind, Ngen, muG, sigmaG, muE, sigmaE, Elim, seedlings, dispersal, viability_deterministic)
-  lapply(list("breeding value (a)" = sim$A, "trait value (z)" = sim$Z), function(M)
+  sim <- simulatorCpp(grid, Ngen, muG, sigmaG, muE, sigmaE, Elim, seedlings, dispersal, viability_deterministic)
+  coords <- data.frame(
+    individual = 1:(grid*grid),
+    X = rep(1:grid, each = grid),
+    Y = rep(1:grid, grid)
+  )
+  lapply(list("ecotype" = sim$E, 
+              "genotype" = sim$A, 
+              "phenotype" = sim$Z), 
+         function(M)
     melt(M) %>% 
     rename(generation = Var1, individual = Var2)) %>% 
     bind_rows(.id = "var") %>% 
-    left_join(data.frame(individual = 1:length(sim$E), environment = sim$E), by = "individual")
+    left_join(coords, by = "individual")
 }