+++ src/product_environmental_matrix.h
@@ -1,0 +1,10 @@
+#ifndef PRODUCT_ENVIRONMENTAL_MATRIX_H
+#define PRODUCT_ENVIRONMENTAL_MATRIX_H
+
+#include <Rcpp.h>
+Rcpp::NumericMatrix product_environmental_matrix(
+    double gradientlim,
+    int length
+);
+#endif
+