+++ NEWS.md
@@ -1,3 +1,3 @@
 # ecoevosimulator 1.0.9001
-
-* 
+* 2D "fixed" simulator
+* mother and father dispersal